Defer shortcut setup to MainActivity (#215)
Shortcuts were being set up (and torn down) in `Application.onCreate()`, which runs in every process and requires eagerly reading `UiStore`. They were also being explicitly removed when the user hid the app icon.

## Changes

- **`MainApplication`**: Remove `setupShortcuts()` and its `uiStore` dependency entirely
- **`MainActivity`**: Call `setupShortcuts()` in `onCreate()`; bail early if `hideAppIcon` is set — no removal needed
- **`AppSettingsActivity`**: Drop `ShortcutManagerCompat.removeAllDynamicShortcuts()` from `onHideIconChange()` — disabling the launcher alias is sufficient; existing shortcuts don't need to be purged

---------
