+++ app/src/main/java/com/github/kr328/clash/AppSettingsActivity.kt
@@ -2,7 +2,6 @@ package com.github.kr328.clash
 
 import android.content.ComponentName
 import android.content.pm.PackageManager
-import androidx.core.content.pm.ShortcutManagerCompat
 import com.github.kr328.clash.common.util.componentName
 import com.github.kr328.clash.design.AppSettingsDesign
 import com.github.kr328.clash.design.model.Behavior
@@ -75,9 +74,5 @@ class AppSettingsActivity : BaseActivity<AppSettingsDesign>(), Behavior {
             newState,
             PackageManager.DONT_KILL_APP
         )
-        if (hide) {
-            // Prevent launcher activity not found.
-            ShortcutManagerCompat.removeAllDynamicShortcuts(this)
-        }
     }
 }
